refactor SAR transformer, add docstrings
